requirements.txt
Flask
Flask-Cors
python-dotenv
requests
fpdf